+++ genja-core/src/lib.rs
@@ -69,6 +69,8 @@ pub use ::async_trait::async_trait;
 #[doc(hidden)]
 pub use ::inventory_crate as __inventory;
 #[doc(hidden)]
+pub use ::schemars as __schemars;
+#[doc(hidden)]
 pub use ::serde as __serde;
 #[doc(hidden)]
 pub use ::serde_json as __serde_json;